refactors code into separate file extensions
